Moved TurnOn/Off Intents to component
